Add files via upload
This code connects an ESP8266 module to a Wi-Fi network and receives sensor data (humidity, temperature, and soil moisture) from an Arduino. The data is parsed from a string and validated. If the data is correct, it is sent to a remote server every 5 seconds using an HTTP POST request.